+++ packages/eslint-plugin-template/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 18.1.0 (2024-07-01)
+
+### 🚀 Features
+
+- **eslint-plugin:** [no-call-expression] add allowPrefix and allowSuffix
+
+### 🩹 Fixes
+
+- update typescript-eslint packages to v8.0.0-alpha.37
+
+- **eslint-plugin-template:** [prefer-self-closing-tags] always ignore index.html files
+
+- **eslint-plugin-template:** [prefer-self-closing-tags] support ng-content with fallback content
+
+### ❤️ Thank You
+
+- Christian Svensson
+- Daniel Kimmich
+- Dave
+- Martijn van der Meij
+- Maximilian Main
+
 ## 18.0.1 (2024-05-30)
 
 ### 🩹 Fixes